FIX: Map jump bug, leaderboard misname
Reintroducing the fix for the bug in tween.js (issue 677) that causes
the map to jump at the end of the tween.

Leaderboard name was used twice in the code base, once for the modal at
the beginning of games, and also for the rotating display in the bottom
left. I've removed the modal at the beggining of the game as its data is
stale and not updated yet. I've renamed the bottom left to
'rotatingDisplay' and the bottom right to 'leaderboard'.